Move new browser unittests to test_example_browser_args.py
Per @mzamoramora-nvidia's review, restore newton/tests/test_example_browser.py
to its original smoke-test contents and put the new auto-discovered
regression tests in newton/tests/test_example_browser_args.py instead.
This avoids touching an existing file just to free up its name.
